# bm_wx_openapi_v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/upload/": {
      "post": {
        "summary": "Upload Files",
        "operationId": "upload_files_upload__post",
        "parameters": [
          {
            "name": "auto_drop_collection_after_search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Drop collection after search",
              "default": true,
              "title": "Auto Drop Collection After Search"
            },
            "description": "Drop collection after search"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_files_upload__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/llm_watsonx_answer/": {
      "post": {
        "summary": "Llm Watsonx Answer",
        "operationId": "llm_watsonx_answer_llm_watsonx_answer__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Hybrid Search Response"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/hybrid_search_by_collection/": {
      "post": {
        "summary": "Hybrid Search By Collection",
        "operationId": "hybrid_search_by_collection_hybrid_search_by_collection__post",
        "parameters": [
          {
            "name": "collection_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_hybrid_search_by_collection_hybrid_search_by_collection__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drop_collection/": {
      "post": {
        "summary": "Drop Collection By Name",
        "operationId": "drop_collection_by_name_drop_collection__post",
        "parameters": [
          {
            "name": "collection_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drop_all_collections/": {
      "post": {
        "summary": "Drop All Collections",
        "operationId": "drop_all_collections_drop_all_collections__post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_hybrid_search_by_collection_hybrid_search_by_collection__post": {
        "properties": {
          "description_input": {
            "type": "string",
            "title": "Description Input",
            "description": "Description JSON input"
          }
        },
        "type": "object",
        "required": [
          "description_input"
        ],
        "title": "Body_hybrid_search_by_collection_hybrid_search_by_collection__post"
      },
      "Body_upload_files_upload__post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files",
            "description": "Upload PDF files"
          },
          "description_input": {
            "type": "string",
            "title": "Description Input",
            "description": "Description JSON input"
          }
        },
        "type": "object",
        "required": [
          "files",
          "description_input"
        ],
        "title": "Body_upload_files_upload__post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ResponseItem": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "relevant_passages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Relevant Passages"
          },
          "llm_answer": {
            "additionalProperties": true,
            "type": "object",
            "title": "Llm Answer"
          }
        },
        "type": "object",
        "required": [
          "description",
          "relevant_passages",
          "llm_answer"
        ],
        "title": "ResponseItem"
      },
      "UploadResponse": {
        "properties": {
          "collection_name": {
            "type": "string",
            "title": "Collection Name"
          },
          "response": {
            "items": {
              "$ref": "#/components/schemas/ResponseItem"
            },
            "type": "array",
            "title": "Response"
          }
        },
        "type": "object",
        "required": [
          "collection_name",
          "response"
        ],
        "title": "UploadResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}